+++ src/clients/avalonstar/canvases/index.js
@@ -1,1 +1,2 @@
 export { default as Activity } from './Activity';
+export { default as Break } from './Break';
