+++ src/containers/Displays/index.js
@@ -1,4 +1,3 @@
 export { default as Activity } from './Activity';
-export { default as Break } from './Break';
 export { default as Intertitle } from './Intertitle';
 export { default as Speedrunning } from './Speedrunning';
